// packages/platform/src/types/mt5/trade/deal.type.ts
export interface Mt5Deal {
  print: string;
  deal: number;
  externalID: string;
  login: number;
  dealer: number;
  order: number;
  action: 'DEAL_BUY' | 'DEAL_SELL';
  entry: 'ENTRY_IN' | 'ENTRY_OUT';
  digits: number;
  digitsCurrency: number;
  contractSize: number;
  time: number;
  symbol: string;
  price: number;
  volume: number;
  profit: number;
  storage: number;
  commission: number;
  rateProfit: number;
  rateMargin: number;
  expertID: number;
  positionID: number;
  comment: string;
  profitRaw: number;
  pricePosition: number;
  volumeClosed: number;
  tickValue: number;
  tickSize: number;
  flags: number;
  timeMsc: number;
  reason: 'DEAL_REASON_CLIENT';
  gateway: string;
  priceGateway: number;
  modificationFlags: 'MODIFY_FLAGS_ADMIN';
  priceSL: number;
  priceTP: number;
  volumeExt: number;
  volumeClosedExt: number;
  fee: number;
  value: number;
  volumeLots: number;
  volumeClosedLots: number;
}